requirements.txt
tqdm
argparse
joblib
opencv-python
onnxruntime-gpu
numpy
pillow
coremltools
uvicorn
fastapi
python-multipart
gunicorn